Move the implementation to flow/terminal
